pyproject.toml
[project]
name = "project-dashboard-api"
version = "0.1.0"
description = "Project management and profiles dashboard backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "asyncpg>=0.29.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.2.1",
    "pyjwt>=2.8.0",
    "pwdlib[bcrypt]>=0.2.0",
    "aioboto3>=12.3.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "httpx>=0.27.0",
    "locust>=2.43.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
